Display remote signatures in file list
