+++ web/src/i18n/locales/default.en-US.json
@@ -88,6 +88,7 @@
     "file.attr.fingerprint": "Fingerprint",
     "file.attr.quality": "Quality",
     "file.attr.duration": "Duration",
+    "file.attr.description": "Description",
     "value.time.seconds": "sec",
     "value.time.minutes": "min",
     "value.time.hours": "hours",
@@ -168,6 +169,7 @@
     "file.match": "Match",
     "file.one": "File",
     "file.many": "Files",
+    "file.remote": "Remote Fingerprint",
     "file.remote.warning": "The owner locked the video metadata",
     "distance.score": "Match score: {value}",
     "filter.title": "Filters",
